+++ changelog.md
@@ -1,5 +1,6 @@
 ### 2.7.4 (2025-06-29)
 - added stepStation and trigStation modules
+- added sampleDelay module
 - added cvMeter module
 - added randMod7 and randMod7compact modules
 - stepSeq8x / trigSeq8x: adjusted RECL and STOR buttons placement